% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaghi.R
\name{binaghi}
\alias{binaghi}
\title{Thematic map accuracy for fuzzy classification/reference data under simple random sampling.}
\usage{
binaghi(r, m, margins = TRUE)
}
\arguments{
\item{r}{dataframe or matrix. Reference data (rows relate to pixels/points and columns to class membership degrees). The columns must be named (see details).}

\item{m}{dataframe or matrix. Membership degrees of a soft classification (rows relate to pixels/points and columns to class membership degrees). The columns must be named (see details).}

\item{margins}{logical. If FALSE, the error matrix produced includes no margins (sum of the 
rows and columns).}
}
\value{
A list with the estimates and matrices, and a csv if defined.
\item{OA}{overall accuracy}
\item{UA}{user's accuracy}
\item{PA}{producer's accuracy}
\item{matrix}{fuzzy confusion error. Rows and columns represent map and reference class labels, respectively.}
\item{numeric}{index of fuzziness of \code{r}}
\item{numeric}{index of fuzziness of \code{m}}
}
\description{
Implements the estimators described in Binaghi \emph{et al}. (1999) for overall accuracy, producer's accuracy, user's accuracy, and index of fuzziness. Includes \emph{no} precision estimates.
}
\details{
The columns of both arguments must be named to explicitly and clearly identify the class that each column refers to. The columns may be sorted differently, and the order of \code{r} will be used for displaying the results.
}
\examples{
## Example of table 2 in Binaghi et al. (1999), p. 940.
# case a (perfect matching):
r<-matrix(c(0.4,0.4,0.4), ncol = 3)
m<-matrix(c(0.4,0.4,0.4), ncol = 3)
colnames(r) <- colnames(m) <- c("q1", "q2", "q3")
binaghi(r,m) # compare to paper

# case b (under-estimation)
r<-matrix(c(0.4,0.4,0.4), ncol = 3)
m<-matrix(c(0.2,0.4,0.4), ncol = 3)
colnames(r) <- colnames(m) <- c("q1", "q2", "q3")
binaghi(r,m) # compare to paper

# case c (over-estimation):
r<-matrix(c(0.4,0.4,0.4), ncol = 3)
m<-matrix(c(0.6,0.4,0.4), ncol = 3)
colnames(r) <- colnames(m) <- c("q1", "q2", "q3")
binaghi(r,m) # compare to paper

## Example of table 3 in Binaghi et al. (1999), p. 941.
# case a (perfect matching):
r<-matrix(c(0.7,0.2,0.1), ncol = 3)
m<-matrix(c(0.7,0.2,0.1), ncol = 3)
colnames(r) <- colnames(m) <- c("q1", "q2", "q3")
binaghi(r,m) # compare to paper

# case b (under-estimation)
r<-matrix(c(0.7,0.2,0.1), ncol = 3)
m<-matrix(c(0.6,0.3,0.1), ncol = 3)
colnames(r) <- colnames(m) <- c("q1", "q2", "q3")
binaghi(r,m) # compare to paper
# (note the error in the paper: UA and PA are swiped)

# case c (over-estimation):
r<-matrix(c(0.7,0.2,0.1), ncol = 3)
m<-matrix(c(0.8,0.1,0.1), ncol = 3)
colnames(r) <- colnames(m) <- c("q1", "q2", "q3")
e<-binaghi(r,m) # compare to paper
# (note the error in the paper: UA and PA are swiped)

## additional example
#m<-readRDS("D:/mypapers/11-mapaccuracy/data/fcla.rds")
#r<-readRDS("D:/mypapers/11-mapaccuracy/data/fref.rds")
#data(fcla)
#data(fref)
#binaghi(fref,fcla)
}
\references{
Binaghi, E.; Brivio, P. A.; Ghezzi, P.; Rampini, A. (1999). \href{https://doi.org/10.1016/S0167-8655(99)00061-6}{A fuzzy set-based accuracy assessment of soft classification}. \emph{Pattern Recognit. Lett.}, 20, 935-948.
}
\seealso{
\code{\link{stehman2007}}
}
\author{
Hugo Costa
}
