% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapAccuracy.R
\docType{methods}
\name{getClassAreas}
\alias{getClassAreas}
\alias{getClassAreas-generic}
\alias{getClassAreas,MapAccuracy-method}
\title{Returns the object's class areas}
\usage{
getClassAreas(object)

\S4method{getClassAreas}{MapAccuracy}(object)
}
\arguments{
\item{object}{A MapAccuracy object}
}
\description{
Returns the object's class areas
}
