% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapAccuracy.R
\docType{methods}
\name{getConfusionMatrix}
\alias{getConfusionMatrix}
\alias{getConfusionMatrix-generic}
\alias{getConfusionMatrix,MapAccuracy-method}
\title{Returns the object's confusion matrix}
\usage{
getConfusionMatrix(object)

\S4method{getConfusionMatrix}{MapAccuracy}(object)
}
\arguments{
\item{object}{A MapAccuracy object}
}
\description{
Returns the object's confusion matrix
}
