% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{internal.mapaccuracy}
\alias{internal.mapaccuracy}
\alias{.check_classes}
\alias{.check_factors}
\alias{.check_labels}
\alias{.check_length}
\title{Internal functions.}
\usage{
.check_classes(x, ...)

.check_factors(x, ...)

.check_labels(x, ...)

.check_length(...)
}
\arguments{
\item{x}{expected objects (character vector)}

\item{...}{arguments to be checked.}
}
\value{
Error message if the test fails. Nothing if the test passes.
}
\description{
Internal functions to be called from within another functions.
}
\details{
\code{.check_classes} checks (tests) if the arguments passed to functions comply with their expected classes.

\code{.check_labels} checks (tests) if the character arguments passed to functions comply with their expected content.

\code{.check_length} checks (tests) if the length of objects match.
}
\keyword{internal}
