% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapAccuracy.R
\docType{class}
\name{MapAccuracy}
\alias{MapAccuracy}
\alias{MapAccuracy-class}
\alias{mapAccuracy}
\title{The MAPACCURACY class}
\usage{
mapAccuracy(confusion_matrix, class_areas)
}
\arguments{
\item{confusion_matrix}{A confusion matrix where the rows contain the results of the classification and the columns the reference data.}

\item{class_areas}{The area of each class in the reference map.}
}
\description{
Use this class for computing the accuracy of a classification using reference map

Creates a MapAccuracy object
}
\note{
No notes
}
\section{Slots }{

\describe{
\item{\code{confusion_matrix}:}{Object of class \code{"matrix"}. A confusion matrix where the rows contain the results of the classification and the columns the reference data.}
\item{\code{class_areas}:}{Object of class \code{"vector"}, representing the area of each class in the reference map.}
}
}

\author{
Alber Sanchez
}
