% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapAccuracy.R
\docType{methods}
\name{setClassAreas}
\alias{setClassAreas}
\alias{setClassAreas,MapAccuracy-method}
\title{Sets the object's class areas}
\usage{
setClassAreas(object, class_areas)

\S4method{setClassAreas}{MapAccuracy}(object, class_areas)
}
\arguments{
\item{object}{A MapAccuracy object}

\item{class_areas}{The area of each class in the map}
}
\description{
Sets the object's class areas
}
