% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapAccuracy.R
\docType{methods}
\name{setConfusionMatrix}
\alias{setConfusionMatrix}
\alias{setConfusionMatrix,MapAccuracy-method}
\title{Sets the object's confusion matrix}
\usage{
setConfusionMatrix(object, confusion_matrix)

\S4method{setConfusionMatrix}{MapAccuracy}(object, confusion_matrix)
}
\arguments{
\item{object}{A MapAccuracy object}

\item{confusion_matrix}{A confusion matrix}
}
\description{
Sets the object's confusion matrix
}
