% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stehman2007.R
\name{stehman2007}
\alias{stehman2007}
\title{Thematic map accuracy for fuzzy classification/reference data under stratified random sampling.}
\usage{
stehman2007(r, m, Nh, old = TRUE)
}
\arguments{
\item{r}{dataframe or matrix. Reference data (rows relate to pixels/points and columns to class membership degrees). The columns must be named (see details).}

\item{m}{dataframe or matrix. Membership degrees of a soft classification (rows relate to pixels/points and columns to class membership degrees). The columns must be named (see details).}

\item{Nh}{numeric vector. Mapped area (or pixel counts) of the classes. It must be named (see details).}

\item{old}{use the old code?}
}
\value{
A list with the estimates and error matrix.
\item{OA}{overall accuracy}
\item{UA}{user's accuracy}
\item{PA}{producer's accuracy}
\item{SEoa}{standard error of OA}
\item{SEua}{standard error of UA}
\item{SEpa}{standard error of PA}
\item{matrix}{confusion error (proportion). Rows and columns represent map and reference class labels, respectively}
}
\description{
Implements the estimators described in Stehman \emph{et al}. (2007) for overall accuracy, producer's accuracy, and user's accuracy. Includes precision estimates.
}
\details{
The columns of both \code{r} and \code{m} must be named to explicitly and clearly identify the class that each column refers to.
Columns may be sorted differently.
Argument \code{Nh} must also be named to identify the class that each area refers to.
The order of \code{Nh} will be used for displaying the results.
Strata area (or pixel counts) is based on the map's spatial unit (e.g. pixel) maximum membership value.
}
\examples{
# Table A1, page 170 of Stehman et al. (2007)
r<-c(0.5, 0.2, 0.3, # stratum 1, pixel 1
        0.4, 0.2, 0.4, # stratum 1, pixel 2
        0.5, 0.2, 0.3, # stratum 1, pixel 3
        0.8, 0.2, 0.0, # stratum 1, pixel 4
        0.4, 0.6, 0.0, # stratum 2, pixel 1
        0.2, 0.6, 0.2, # stratum 2, pixel 2
        0.6, 0.1, 0.3, # stratum 2, pixel 3
        0.5, 0.5, 0.0, # stratum 2, pixel 4
        0.2, 0.1, 0.7, # stratum 3, pixel 1
        0.6, 0.4, 0.0, # stratum 3, pixel 2
        0.0, 0.3, 0.7, # stratum 3, pixel 3
        0.9, 0.0, 0.1) # stratum 3, pixel 4
r<-matrix(r,ncol=3,byrow = TRUE)

m<-c(0.6, 0.3, 0.1, # stratum 1, pixel 1
        0.4, 0.3, 0.3, # stratum 1, pixel 2
        0.6, 0.0, 0.4, # stratum 1, pixel 3
        1.0, 0.0, 0.0, # stratum 1, pixel 4
        0.1, 0.7, 0.2, # stratum 2, pixel 1
        0.3, 0.4, 0.3, # stratum 2, pixel 2
        0.3, 0.5, 0.2, # stratum 2, pixel 3
        0.0, 1.0, 0.0, # stratum 2, pixel 4
        0.1, 0.1, 0.8, # stratum 3, pixel 1
        0.3, 0.3, 0.4, # stratum 3, pixel 2
        0.0, 0.3, 0.7, # stratum 3, pixel 3
        0.2, 0.3, 0.5) # stratum 3, pixel 4
m<-matrix(m,ncol = 3, byrow = TRUE)
colnames(r) <- colnames(m) <- c("N1", "N2", "N3")
Nh<-c("N1"=500, "N2"=300, "N3"=200)
e<-stehman2007(r, m, Nh)

# compare to paper (differences due to number rounding):
e$OA          # eq. 170
e$SEoa^2      # eq. 170
e$SEoa        # eq. 171
e$UA[2]       # eq. 171
e$SEua[2]^2   # eq. 172 ######################## ERRO no meu codigo: should be 0.0826
e$SEua[2]     # eq. 172 ######################## ERRO no meu codigo: should be 0.287
e$PA[3]       # eq. 172
e$SEpa[3]^2   # eq. 172
e$SEpa[3]     # eq. 173
}
\references{
Stehman, S. V.; Arora, M. K.; Kasetkasem, T.; Varshney, P. K (2007). \href{https://doi.org/10.14358/PERS.73.2.165}{Estimation of fuzzy error matrix accuracy measures under stratified random sampling}. \emph{Photogramm. Eng. Remote Sensing}, 73, 165–173.
}
\author{
Hugo Costa
}
