% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stehman2014.R
\name{stehman2014}
\alias{stehman2014}
\title{Thematic map accuracy and area under stratified random sampling when the strata differ from the map classes.}
\usage{
stehman2014(s, r, m, Nh_strata, margins = TRUE)
}
\arguments{
\item{s}{character vector. Strata class labels. The object will be coarsed to factor.}

\item{r}{character vector. Reference class labels. The object will be coarsed to factor.}

\item{m}{character vector. Map class labels. The object will be coarsed to factor.}

\item{Nh_strata}{numeric vector. Area of the classes in the strata. It must be named (see details).}

\item{margins}{logical. If FALSE, the error matrix produced includes no margins (sum of the rows and columns).}
}
\value{
A list with the estimates and error matrix.
\item{OA}{overall accuracy}
\item{UA}{user's accuracy}
\item{PA}{producer's accuracy}
\item{area}{area}
\item{SEoa}{standard error of OA}
\item{SEua}{standard error of UA}
\item{SEpa}{standard error of PA}
\item{SEa}{standard error of area}
\item{matrix}{confusion error (area proportion). Rows and columns represent map and reference class labels, respectively}
}
\description{
Implements the estimators described in Stehman (2014) for overall accuracy, producer's accuracy, user's accuracy, and area. Includes precision estimates.
}
\details{
Argument \code{Nh_strata} must be named to explicitly and clearly identify the class that each area refers to.
The order of \code{Nh} will be used for displaying the results.

The labels included in \code{s}, \code{r} and \code{m} must match among them and the names of \code{Nh_strata}.

The estimates of area are given in the same unit of the mapped areas (if \code{Nh_strata} express hectares, area estimates express hectares as well).
}
\examples{
# Numerical example in Stheman (2014)
s<-c(rep("A",10), rep("B",10), rep("C",10), rep("D",10))
m<-c(rep("A",7), rep("B",3), "A", rep("B",11), rep("C",6), "B", "B", rep("D",10))
r<-c(rep("A",5), "C", "B", "A", "B", "C", "A", rep("B",5),
              "A", "A", "B", "B", rep("C",5), "D", "D", "B", "B", "A", rep("D",7), "C", "C", "B")
Nh_strata<-c("A"=40000, "B"=30000, "C"=20000, "D"=10000)
e<-stehman2014(s, r, m, Nh_strata)
(e$area/sum(Nh_strata))[1] # Proportion of area of class A (compare with paper in p. 4932)
(e$area/sum(Nh_strata))[3] # Proportion of area of class C (p. 4932)
e$OA                       # Overall accuracy (p. 4932)
e$UA[2]                    # User's accuracy of class B (compare with paper in p. 4934)
e$PA[2]                    # Producer's accuracy of class B (p. 4934)
e$matrix[2,3]              # Cell (2, 3) of the error matrix (p. 4935)
e$SEa[1]/Nh_strata[1]      # Standard error (SE) for proportion of area of class A (p. 4935)
e$SEa[3]/Nh_strata[3]      # Standard error (SE) for proportion of area of class C (p. 4935)
e$SEoa                     # SE for overall accuracy (p. 4936)
e$SEua[2]                  # SE for user's accuracy of class B (p. 4936)
e$SEpa[2]                  # SE for producer's accuracy of class B (p. 4936)

# change class order
stehman2014(s, m, r, Nh_strata[c(4,2,1,3)], margins=FALSE)
}
\references{
Stehman, S. V. (2014).\href{https://doi.org/10.1080/01431161.2014.930207}{Estimating area and map accuracy for stratified random sampling when the strata are different from the map classes}. \emph{Int. J. Remote Sens.}, 35, 4923-4939.
}
\author{
Hugo Costa
}
